refactor: model -> domain 패키지 수정
