+++ src/main/java/lotterymachine/domain/LotteryTicket.java
@@ -1,4 +1,4 @@
-package lotterymachine.model;
+package lotterymachine.domain;
 
 import java.util.Collections;
 import java.util.List;
