+++ src/main/java/lotterymachine/domain/LotteryTickets.java
@@ -1,4 +1,4 @@
-package lotterymachine.model;
+package lotterymachine.domain;
 
 import java.util.*;
