+++ src/test/java/lotterymachine/domain/LotteryTicketTest.java
@@ -1,4 +1,4 @@
-package lotterymachine.model;
+package lotterymachine.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
