+++ src/test/java/lotterymachine/utils/LotteryRuleTest.java
@@ -2,6 +2,7 @@ package lotterymachine.utils;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import lotterymachine.domain.LotteryRule;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
